Issue-14 - Infinite reminder comments
The condition for diff > 24*Hours || diff < -24*Hours breaks when
crossing day bounderies. The date of the posted reminder is before the
day that it parses as deadline and thus it doesn't match when it's
searching for it at line 207.

When we ensure we are only looking for reminders for the same day we are
not allowing it to cross the day and will always match against the
current date

Signed-off-by: Makpoc <makpoc@gmail.com>